requirements.txt
pyopengl
pygame
SimpleGUICS2Pygame
glfw
numpy
# serial